requirements.txt
# Core (required)
fastapi==0.115.0
uvicorn==0.30.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
requests==2.32.0
numpy==1.26.4
pydantic==2.9.0

# Optional: only needed for generate_q_variants.py
anthropic==0.40.0
